requirements.txt
python-dotenv==1.0.1
langchain==0.1.5
langchain-community==0.0.18
langchain-core==0.1.19
langchain-google-genai==0.0.6
langsmith==0.0.86
docx2txt==0.8
boto3==1.34.74
botocore==1.34.74
streamlit==1.41.1
altair<5
setuptools